Adds `id` attributes to all `h2` elements (#47)
I wanted to send someone a link to the list of projects, but there wasn't a way to link to just that section of the page, because there are no `id` attributes for the `h2` elements. Adding these will make it easier to link to specific sections. This could also help with adding a table of contents to the top of the page in the future.